pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "nudemo"
version = "0.1.0"
description = "nuScenes mini dataset generation and benchmarking pipeline."
readme = "README.md"
requires-python = ">=3.12,<3.13"
license = { text = "MIT" }
authors = [
  { name = "cosmobean", email = "cosmobean.dev@gmail.com" },
]
dependencies = [
  "numpy>=1.26,<2",
  "Pillow>=11.0,<12",
]

[project.optional-dependencies]
pipeline = [
  "psycopg[binary]>=3.2,<4",
  "minio>=7.2,<8",
  "redis>=5.2,<6",
  "pyarrow>=19.0,<20",
  "lance>=1.2.1,<2",
  "webdataset>=1.0.2,<2",
  "confluent-kafka>=2.13,<3",
]
training = [
  "torch>=2.6,<3",
  "torchvision>=0.21,<1",
]
reporting = [
  "pandas>=2.2,<3",
  "plotly>=6.0,<7",
  "dash>=2.18,<3",
]
dataset = [
  "nuscenes-devkit>=1.2.0,<2",
  "tqdm>=4.67,<5",
]
dev = [
  "pytest>=8.3,<9",
  "ruff>=0.11,<0.12",
]

[project.scripts]
nudemo = "nudemo.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/nudemo"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["E402", "E501"]
